DSD-0004 moved types around
