+++ packages/Core/src/Interfaces/IResourceModule.ts
@@ -1,5 +1,4 @@
-
-import { TResourceHandle, TConfig } from './GeneralInterfaces';
+import { TResourceHandle, TConfig } from '../types';
 
 /**
  * interface for engine modules that have CRUD interfaces for their resources
